Fix IS_PIN_INTERRUPT in Boards.h
Interrupt numbers start at zero. The function `digitalPinToInterrupt()` correctly returns interrupt number `0` for `pin 2`, but this makes `IS_PIN_INTERRUPT(2)` false. I found the bug because the `FirmataEncoder` library checks `IS_PIN_INTERRUPT()` to determine whether a digital pin can be used as an encoder. It was incorrectly reporting pin 2 on an Arduino as incapable of being used as an encoder.

The correct way is to check whether `digitalPinToInterrupt()` returns a valid interrupt other than `NOT_AN_INTERRUPT` (=`-1`).